scraper/requirements.txt
pymongo
requests
pandas
python-dotenv
